Time: 55 ms (78.5%), Space: 108 MB (63.5%) - LeetHub

+++ 3867-sum-of-gcd-of-formed-pairs/3867-sum-of-gcd-of-formed-pairs.java
@@ -15,7 +15,7 @@ class Solution {
         int left = 0, right = n-1;
         long sum = 0;
         while(left < right){
-            sum += (long) GCD(arr[right--], arr[left++]);
+            sum +=  GCD(arr[right--], arr[left++]);
         }
         return sum;
     }
